+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/domain/dto/requests/RequestUserUpdate2.java
@@ -1,0 +1,10 @@
+package io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.requests;
+
+import lombok.Data;
+
+// Lombok
+@Data
+public class RequestUserUpdate2 {
+    private final String zoneId;
+    private final String name;
+}