+++ tech1-framework-b2b-mongodb-security-jwt/src/test/java/io/tech1/framework/b2b/mongodb/security/jwt/resources/BaseSecurityUserResourceTest.java
@@ -4,6 +4,7 @@ import io.tech1.framework.b2b.mongodb.security.jwt.assistants.core.CurrentSessio
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.db.DbUser;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.requests.RequestUserChangePassword1;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.requests.RequestUserUpdate1;
+import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.requests.RequestUserUpdate2;
 import io.tech1.framework.b2b.mongodb.security.jwt.services.BaseUserService;
 import io.tech1.framework.b2b.mongodb.security.jwt.tests.runnerts.AbstractResourcesRunner;
 import io.tech1.framework.b2b.mongodb.security.jwt.validators.BaseUserValidator;
@@ -70,6 +71,24 @@ public class BaseSecurityUserResourceTest extends AbstractResourcesRunner {
         verify(this.currentSessionAssistant).getCurrentDbUser();
         verify(this.baseUserValidator).validateUserUpdateRequest1(eq(currentDbUser), eq(requestUserUpdate1));
         verify(this.baseUserService).updateUser1(eq(requestUserUpdate1));
+    }
+
+    @Test
+    public void update2Test() throws Exception {
+        // Arrange
+        var requestUserUpdate2 = entity(RequestUserUpdate2.class);
+
+        // Act
+        this.mvc.perform(
+                        post("/user/update2")
+                                .content(this.objectMapper.writeValueAsString(requestUserUpdate2))
+                                .contentType(MediaType.APPLICATION_JSON)
+                )
+                .andExpect(status().isOk());
+
+        // Assert
+        verify(this.baseUserValidator).validateUserUpdateRequest2(eq(requestUserUpdate2));
+        verify(this.baseUserService).updateUser2(eq(requestUserUpdate2));
     }
 
     @Test