+++ tech1-framework-b2b-mongodb-security-jwt/src/test/java/io/tech1/framework/b2b/mongodb/security/jwt/services/impl/BaseUserServiceImplTest.java
@@ -4,6 +4,7 @@ import io.tech1.framework.b2b.mongodb.security.jwt.assistants.core.CurrentSessio
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.db.DbUser;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.requests.RequestUserChangePassword1;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.requests.RequestUserUpdate1;
+import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.requests.RequestUserUpdate2;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.jwt.JwtUser;
 import io.tech1.framework.b2b.mongodb.security.jwt.repositories.UserRepository;
 import io.tech1.framework.b2b.mongodb.security.jwt.services.BaseUserService;
@@ -106,6 +107,30 @@ public class BaseUserServiceImplTest {
         assertThat(dbUserAC.getValue().getZoneId()).isEqualTo(ZoneId.of(requestUserUpdate1.getZoneId()));
         assertThat(dbUserAC.getValue().getName()).isEqualTo(requestUserUpdate1.getName());
         assertThat(dbUserAC.getValue().getEmail()).isEqualTo(requestUserUpdate1.getEmail());
+        // WARNING: no verifications on static SecurityContextHolder
+    }
+
+    @Test
+    public void updateUser2Test() {
+        // Arrange
+        var requestUserUpdate2 = new RequestUserUpdate2(
+                randomZoneId().getId(),
+                randomString()
+        );
+        var currentJwtUser = entity(JwtUser.class);
+        when(this.currentSessionAssistant.getCurrentJwtUser()).thenReturn(currentJwtUser);
+        var dbUserAC = ArgumentCaptor.forClass(DbUser.class);
+
+        // Act
+        this.componentUnderTest.updateUser2(requestUserUpdate2);
+
+        // Assert
+        verify(this.currentSessionAssistant).getCurrentJwtUser();
+        verify(this.userRepository).save(dbUserAC.capture());
+        assertThat(dbUserAC.getValue().getUsername()).isEqualTo(currentJwtUser.getDbUser().getUsername());
+        assertThat(dbUserAC.getValue().getZoneId()).isEqualTo(ZoneId.of(requestUserUpdate2.getZoneId()));
+        assertThat(dbUserAC.getValue().getName()).isEqualTo(requestUserUpdate2.getName());
+        assertThat(dbUserAC.getValue().getEmail()).isEqualTo(currentJwtUser.getDbUser().getEmail());
         // WARNING: no verifications on static SecurityContextHolder
     }